+++ mrv2/docs/HISTORY.md
@@ -65,6 +65,13 @@ v1.0.9
 - Fixed gcc version information in mrv2's About Window.  It was showing the
   __GLIBCXX__ instead.
 - Darby simplified and fixed the A/B comparison sometimes getting out of sync.
+- Improved the look of HorSlider widget.  Now the X to reset to defaults is
+  a tad to the right.
+- Fixed selection of Wipe mode which would not work after scrolling in the
+  docker (FLTK bug?).
+- Made Compare options automatically frame the view if View->Auto Frame is
+  on.
+- Improved the playback performance of the viewer a tad.
 
 
 v1.0.8